🧑‍💼️ Switch to Nix and home-manager (#1)
This switches us from a series of spreadsheets (and trying to deal with potentially-different versions of software on different platforms) to a home manager configuration that should remain relatively consistent across systems.

Based on a couple transfers since creating the branch, this seems to work a little better, but some things are clearly still broken. But I think this is in good enough shape at this point that I should stop operating out of the branch full time.